Add size tracking to MemoryQueue and update tests
A size method and attribute have been added to the MemoryQueue structure in messageq.go to track the current size of the queue. This size increases when an item is enqueued and decreases when an item is dequeued. The messageq_test.go file has been updated with tests for the size method and Dequeue operation of an empty queue. The TestConcurrency test has also been updated to check for thread safety of queue size.